# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOFTWORKS IT FARM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Portfolio
  /portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: List portfolio items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: createPortfolioItem
      tags: [portfolio]
      summary: Create portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
  /portfolio/{id}:
    get:
      operationId: getPortfolioItem
      tags: [portfolio]
      summary: Get portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
    put:
      operationId: updatePortfolioItem
      tags: [portfolio]
      summary: Update portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItemBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
    delete:
      operationId: deletePortfolioItem
      tags: [portfolio]
      summary: Delete portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Blog
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    put:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Team
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    put:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SaaS Products
  /saas-products:
    get:
      operationId: listSaasProducts
      tags: [saas]
      summary: List SaaS products
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaasProduct"
    post:
      operationId: createSaasProduct
      tags: [saas]
      summary: Create SaaS product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaasProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaasProduct"
  /saas-products/{id}:
    put:
      operationId: updateSaasProduct
      tags: [saas]
      summary: Update SaaS product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaasProductBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaasProduct"
    delete:
      operationId: deleteSaasProduct
      tags: [saas]
      summary: Delete SaaS product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Employees (HR)
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Clients
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - icon
        - category
        - features
        - isActive
        - createdAt
    CreateServiceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required:
        - title
        - description
        - icon
        - category
        - features
    PortfolioItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        projectUrl:
          type: string
        clientName:
          type: string
        isFeatured:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - technologies
        - imageUrl
        - clientName
        - isFeatured
        - createdAt
    CreatePortfolioItemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        projectUrl:
          type: string
        clientName:
          type: string
        isFeatured:
          type: boolean
      required:
        - title
        - description
        - category
        - technologies
        - imageUrl
        - clientName
...
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        company:
          type: string
        role:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatarUrl:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clientName
        - company
        - content
        - rating
        - isActive
        - createdAt
    CreateTestimonialBody:
      type: object
      properties:
        clientName:
          type: string
        company:
          type: string
        role:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatarUrl:
          type: string
        isActive:
          type: boolean
      required:
        - clientName
        - company
        - content
        - rating
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - role
        - department
        - bio
        - skills
        - isActive
        - createdAt
    CreateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        department:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - role
        - department
        - bio
        - skills
    SaasProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        features:
          type: array
          items:
            type: string
        pricingMonthly:
          type: number
        pricingYearly:
          type: number
        iconUrl:
          type: string
        demoUrl:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - category
        - status
        - features
        - isActive
        - createdAt
    CreateSaasProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        features:
          type: array
          items:
            type: string
        pricingMonthly:
          type: number
        pricingYearly:
          type: number
        iconUrl:
          type: string
        demoUrl:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - description
        - category
        - status
        - features
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        status:
          type: string
        priority:
          type: string
        budget:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        progress:
          type: integer
        technologies:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - clientId
        - clientName
        - status
        - priority
        - budget
        - startDate
        - endDate
        - progress
        - technologies
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        status:
          type: string
        priority:
          type: string
        budget:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        progress:
          type: integer
        technologies:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - clientId
        - clientName
        - status
        - priority
        - budget
        - startDate
        - endDate
        - progress
        - technologies
...