# lib/api-zod/src/generated/types/createSaasProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SOFTWORKS IT FARM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSaasProductBody {
  name: string;
  description: string;
  category: string;
  status: string;
  features: string[];
  pricingMonthly?: number;
  pricingYearly?: number;
  iconUrl?: string;
  demoUrl?: string;
  isActive?: boolean;
}