# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SOFTWORKS IT FARM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  description: string;
  icon: string;
  category: string;
  features: string[];
  isActive: boolean;
  createdAt: string;
}